+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.35]
+- added when(..) and whenHasValue(...) to Field
+- added initial value to Field factories
+- Added BitLabel
+
 ## [0.0.34]
 - CalendarSlot returns proper time on empty slot